+++ src/Field.hpp
@@ -6,7 +6,7 @@
 constexpr int kFieldWidth = 3;
 constexpr int kFieldHeight = 10;
 
-constexpr float kSquareSize = 0.5;
+constexpr double kSquareSize = 0.5;
 
 // Here lie functions & structures to manage the legalmovesmatrix
