+++ src/PID.cpp
@@ -5,7 +5,7 @@
 
 
 
-struct Error update_error(struct Error err, float value, float setpoint, float delta_s)
+struct Error update_error(struct Error err, double value, double setpoint, double delta_s)
 {
 	// Serial.print("Delta s: ");
 	// Serial.println(delta_s);
@@ -15,7 +15,7 @@ struct Error update_error(struct Error err, float value, float setpoint, float d
 	out.sum_error = err.sum_error * 0.999 + out.error * delta_s;
 	return out;
 }
-float get(struct PID const& pid, struct Error const& error)
+double get(struct PID const& pid, struct Error const& error)
 {
 	return pid.P * error.error + pid.D * error.diff_error + pid.I * error.sum_error;
 }
